Remove check that avoids using holes on finals
I distinctly remembmer adding that check in order to combat some
misaligned transitions, and testing it by hand to ensure that it worked.
Now I'm kicking myself for not having a test suite.

Really, though, what we need is to hook up CV libraries to deal with
some of these questions automatically.